Extract separate tests out of unit/edition/gov_uk_delivery_test
